Move the connection from the homestacks to hometabs
